[Lua]Move vec/mat into a new file
