+++ Src/Module/LuaScript/LuaInterpreter/config.ld
@@ -5,7 +5,7 @@ format='markdown'
 backtick_references=false
 readme='intro.md'
 wrap=true
-file={'./interpreter.cpp','./lua_vessel_mtd.cpp','./lua_xrsound.cpp','./types.lua'}
+file={'./interpreter.cpp','./lua_vessel_mtd.cpp','./lua_xrsound.cpp','lua_vecmat.cpp','./types.lua'}
 dir='out'
 style='!fixed'
 kind_names={topic='Manual',script='Programs'}
